@@ -3,7 +3,6 @@ require 'active_support/inflector'
 require 'set'
 
 module JIRA
-
   # This class provides the basic object <-> REST mapping for all JIRA::Resource subclasses,
   # i.e. the Create, Retrieve, Update, Delete lifecycle methods.
   #
@@ -51,8 +50,8 @@ module JIRA
   #   new_comment = issue.comments.build
   #
   class Base
-    QUERY_PARAMS_FOR_SINGLE_FETCH = Set.new [:expand, :fields]
-    QUERY_PARAMS_FOR_SEARCH = Set.new [:expand, :fields, :startAt, :maxResults]
+    QUERY_PARAMS_FOR_SINGLE_FETCH = Set.new %i[expand fields]
+    QUERY_PARAMS_FOR_SEARCH = Set.new %i[expand fields startAt maxResults]
 
     # A reference to the JIRA::Client used to initialize this resource.
     attr_reader :client
@@ -67,8 +66,8 @@ module JIRA
     # representation of the JSON returned from the JIRA API
     attr_accessor :attrs
 
-    alias :expanded? :expanded
-    alias :deleted? :deleted
+    alias expanded? expanded
+    alias deleted? deleted
 
     def initialize(client, options = {})
       @client   = client
@@ -80,12 +79,12 @@ module JIRA
       # each of them must be passed in to the initializer.
       self.class.belongs_to_relationships.each do |relation|
         if options[relation]
-          instance_variable_set("@#{relation.to_s}", options[relation])
-          instance_variable_set("@#{relation.to_s}_id", options[relation].key_value)
+          instance_variable_set("@#{relation}", options[relation])
+          instance_variable_set("@#{relation}_id", options[relation].key_value)
         elsif options["#{relation}_id".to_sym]
-          instance_variable_set("@#{relation.to_s}_id", options["#{relation}_id".to_sym])
+          instance_variable_set("@#{relation}_id", options["#{relation}_id".to_sym])
         else
-          raise ArgumentError.new("Required option #{relation.inspect} missing") unless options[relation]
+          raise ArgumentError, "Required option #{relation.inspect} missing" unless options[relation]
         end
       end
     end
@@ -95,17 +94,15 @@ module JIRA
     def self.all(client, options = {})
       response = client.get(collection_path(client))
       json = parse_json(response.body)
-      if collection_attributes_are_nested
-        json = json[endpoint_name.pluralize]
-      end
+      json = json[endpoint_name.pluralize] if collection_attributes_are_nested
       json.map do |attrs|
-        self.new(client, {:attrs => attrs}.merge(options))
+        new(client, { attrs: attrs }.merge(options))
       end
     end
 
     # Finds and retrieves a resource with the given ID.
     def self.find(client, key, options = {})
-      instance = self.new(client, options)
+      instance = new(client, options)
       instance.attrs[key_attribute.to_s] = key
       instance.fetch(false, query_params_for_single_fetch(options))
       instance
@@ -114,7 +111,7 @@ module JIRA
     # Builds a new instance of the resource with the given attributes.
     # These attributes will be posted to the JIRA Api if save is called.
     def self.build(client, attrs)
-      self.new(client, :attrs => attrs)
+      new(client, attrs: attrs)
     end
 
     # Returns the name of this resource for use in URL components.
@@ -122,7 +119,7 @@ module JIRA
     #   JIRA::Resource::Issue.endpoint_name
     #     # => issue
     def self.endpoint_name
-      self.name.split('::').last.downcase
+      name.split('::').last.downcase
     end
 
     # Returns the full path for a collection of this resource.
@@ -130,7 +127,7 @@ module JIRA
     #   JIRA::Resource::Issue.collection_path
     #     # => /jira/rest/api/2/issue
     def self.collection_path(client, prefix = '/')
-      client.options[:rest_base_path] + prefix + self.endpoint_name
+      client.options[:rest_base_path] + prefix + endpoint_name
     end
 
     # Returns the singular path for the resource with the given key.
@@ -144,7 +141,7 @@ module JIRA
     #   JIRA::Resource::Comment.singular_path('456','/issue/123/')
     #     # => /jira/rest/api/2/issue/123/comment/456
     def self.singular_path(client, key, prefix = '/')
-      collection_path(client, prefix) + '/' + key
+      collection_path(client, prefix) + '/' + key.to_s
     end
 
     # Returns the attribute name of the attribute used for find.
@@ -200,7 +197,7 @@ module JIRA
       define_method(resource) do
         attribute = maybe_nested_attribute(attribute_key, options[:nested_under])
         return nil unless attribute
-        child_class.new(client, :attrs => attribute)
+        child_class.new(client, attrs: attribute)
       end
     end
 
@@ -248,12 +245,12 @@ module JIRA
     def self.has_many(collection, options = {})
       attribute_key = options[:attribute_key] || collection.to_s
       child_class = options[:class] || ('JIRA::Resource::' + collection.to_s.classify).constantize
-      self_class_basename = self.name.split('::').last.downcase.to_sym
+      self_class_basename = name.split('::').last.downcase.to_sym
       define_method(collection) do
-        child_class_options = {self_class_basename => self}
+        child_class_options = { self_class_basename => self }
         attribute = maybe_nested_attribute(attribute_key, options[:nested_under]) || []
         collection = attribute.map do |child_attributes|
-          child_class.new(client, child_class_options.merge(:attrs => child_attributes))
+          child_class.new(client, child_class_options.merge(attrs: child_attributes))
         end
         HasManyProxy.new(self, child_class, collection)
       end
@@ -290,8 +287,8 @@ module JIRA
     # Checks if method_name is set in the attributes hash
     # and returns true when found, otherwise proxies the
     # call to the superclass.
-    def respond_to?(method_name, include_all=false)
-      if attrs.keys.include? method_name.to_s
+    def respond_to?(method_name, _include_all = false)
+      if attrs.key?(method_name.to_s)
         true
       else
         super(method_name)
@@ -301,8 +298,8 @@ module JIRA
     # Overrides method_missing to check the attribute hash
     # for resources matching method_name and proxies the call
     # to the superclass if no match is found.
-    def method_missing(method_name, *args, &block)
-      if attrs.keys.include? method_name.to_s
+    def method_missing(method_name, *_args)
+      if attrs.key?(method_name.to_s)
         attrs[method_name.to_s]
       else
         super(method_name)
@@ -315,7 +312,7 @@ module JIRA
       @attrs[self.class.key_attribute.to_s]
     end
 
-    def collection_path(prefix = "/")
+    def collection_path(prefix = '/')
       # Just proxy this to the class method
       self.class.collection_path(client, prefix)
     end
@@ -325,9 +322,7 @@ module JIRA
     # issue it returns '/issue'
     def path_component
       path_component = "/#{self.class.endpoint_name}"
-      if key_value
-        path_component += '/' + key_value
-      end
+      path_component += '/' + key_value if key_value
       path_component
     end
 
@@ -346,9 +341,10 @@ module JIRA
     #
     # Accepts an attributes hash of the values to be saved.  Will throw a
     # JIRA::HTTPError if the request fails (response is not HTTP 2xx).
-    def save!(attrs)
+    def save!(attrs, path = nil)
+      path ||= new_record? ? url : patched_url
       http_method = new_record? ? :post : :put
-      response = client.send(http_method, new_record? ? url : patched_url, attrs.to_json)
+      response = client.send(http_method, path, attrs.to_json)
       set_attrs(attrs, false)
       set_attrs_from_response(response)
       @expanded = false
@@ -360,19 +356,18 @@ module JIRA
     #
     # Accepts an attributes hash of the values to be saved. Will return false
     # if the request fails.
-    def save(attrs)
+    def save(attrs, path = url)
       begin
-        save_status = save!(attrs)
+        save_status = save!(attrs, path)
       rescue JIRA::HTTPError => exception
         begin
           set_attrs_from_response(exception.response) # Merge error status generated by JIRA REST API
         rescue JSON::ParserError => parse_exception
-          set_attrs("exception" => {
-                        "class" => exception.response.class.name,
-                        "code" => exception.response.code,
-                        "message" => exception.response.message
-                    }
-          )
+          set_attrs('exception' => {
+                      'class' => exception.response.class.name,
+                      'code' => exception.response.code,
+                      'message' => exception.response.message
+                    })
         end
         # raise exception
         save_status = false
@@ -383,7 +378,7 @@ module JIRA
     # Sets the attributes hash from a HTTPResponse object from JIRA if it is
     # not nil or is not a json response.
     def set_attrs_from_response(response)
-      unless response.body.nil? or response.body.length < 2
+      unless response.body.nil? || (response.body.length < 2)
         json = self.class.parse_json(response.body)
         set_attrs(json)
       end
@@ -393,7 +388,7 @@ module JIRA
     # hash values will be clobbered by the new hash, otherwise the hash will
     # be deeply merged into attrs.  The target paramater is for internal use only
     # and should not be used.
-    def set_attrs(hash, clobber=true, target = nil)
+    def set_attrs(hash, clobber = true, target = nil)
       target ||= @attrs
       if clobber
         target.merge!(hash)
@@ -425,12 +420,12 @@ module JIRA
       prefix = '/'
       unless self.class.belongs_to_relationships.empty?
         prefix = self.class.belongs_to_relationships.inject(prefix) do |prefix_so_far, relationship|
-          prefix_so_far.to_s + relationship.to_s + "/" + self.send("#{relationship.to_s}_id").to_s + '/'
+          prefix_so_far.to_s + relationship.to_s + '/' + send("#{relationship}_id").to_s + '/'
         end
       end
       if @attrs['self']
         the_url = @attrs['self']
-        the_url = the_url.sub(@client.options[:site], '') if @client.options[:site]
+        the_url = the_url.sub(@client.options[:site].chomp('/'), '') if @client.options[:site]
         the_url
       elsif key_value
         self.class.singular_path(client, key_value.to_s, prefix)
@@ -448,7 +443,7 @@ module JIRA
     # [07/Jun/2015:15:17:18 +0400] "PUT /jira/rest/api/2/issue/10111 HTTP/1.1" 204 -
     def patched_url
       result = url
-      return result if result.start_with?('/')
+      return result if result.start_with?('/', 'http')
       "/#{result}"
     end
 
@@ -495,9 +490,15 @@ module JIRA
     end
 
     def url_with_query_params(url, query_params)
-      uri = URI.parse(url)
-      uri.query = uri.query.nil? ? "#{hash_to_query_string query_params}" : "#{uri.query}&#{hash_to_query_string query_params}" unless query_params.empty?
-      uri.to_s
+      self.class.url_with_query_params(url, query_params)
+    end
+
+    def self.url_with_query_params(url, query_params)
+      if !query_params.empty?
+        "#{url}?#{hash_to_query_string query_params}"
+      else
+        url
+      end
     end
 
     def hash_to_query_string(query_params)
@@ -505,19 +506,19 @@ module JIRA
     end
 
     def self.hash_to_query_string(query_params)
-      query_params.map do |k,v|
+      query_params.map do |k, v|
         CGI.escape(k.to_s) + '=' + CGI.escape(v.to_s)
       end.join('&')
     end
 
     def self.query_params_for_single_fetch(options)
-      Hash[options.select do |k,v|
+      Hash[options.select do |k, _v|
         QUERY_PARAMS_FOR_SINGLE_FETCH.include? k
       end]
     end
 
     def self.query_params_for_search(options)
-      Hash[options.select do |k,v|
+      Hash[options.select do |k, _v|
         QUERY_PARAMS_FOR_SEARCH.include? k
       end]
     end
