@@ -5,11 +5,10 @@ require 'uri'
 
 module JIRA
   class HttpClient < RequestClient
-
     DEFAULT_OPTIONS = {
-      :username           => '',
-      :password           => ''
-    }
+      username: nil,
+      password: nil
+    }.freeze
 
     attr_reader :options
 
@@ -19,23 +18,26 @@ module JIRA
     end
 
     def make_cookie_auth_request
-      body = { :username => @options[:username], :password => @options[:password] }.to_json
+      body = { username: @options[:username].to_s, password: @options[:password].to_s }.to_json
       @options.delete(:username)
       @options.delete(:password)
-      make_request(:post, @options[:context_path] + '/rest/auth/1/session', body, {'Content-Type' => 'application/json'})
+      make_request(:post, @options[:context_path] + '/rest/auth/1/session', body, 'Content-Type' => 'application/json')
     end
 
-    def make_request(http_method, url, body='', headers={})
+    def make_request(http_method, url, body = '', headers = {})
       # When a proxy is enabled, Net::HTTP expects that the request path omits the domain name
       path = request_path(url)
       request = Net::HTTP.const_get(http_method.to_s.capitalize).new(path, headers)
       request.body = body unless body.nil?
-      add_cookies(request) if options[:use_cookies]
-      request.basic_auth(@options[:username], @options[:password]) if @options[:username] && @options[:password]
-      response = basic_auth_http_conn.request(request)
-      @authenticated = response.is_a? Net::HTTPOK
-      store_cookies(response) if options[:use_cookies]
-      response
+
+      execute_request(request)
+    end
+
+    def make_multipart_request(url, body, headers = {})
+      path = request_path(url)
+      request = Net::HTTP::Post::Multipart.new(path, body, headers)
+
+      execute_request(request)
     end
 
     def basic_auth_http_conn
@@ -43,20 +45,26 @@ module JIRA
     end
 
     def http_conn(uri)
-      if @options[:proxy_address]
-          http_class = Net::HTTP::Proxy(@options[:proxy_address], @options[:proxy_port] ? @options[:proxy_port] : 80)
-      else
-          http_class = Net::HTTP
-      end
-      http_conn = http_class.new(uri.host, uri.port)
+      http_conn =
+        if @options[:proxy_address]
+          Net::HTTP.new(uri.host, uri.port, @options[:proxy_address], @options[:proxy_port] || 80, @options[:proxy_username], @options[:proxy_password])
+        else
+          Net::HTTP.new(uri.host, uri.port)
+        end
       http_conn.use_ssl = @options[:use_ssl]
+      if @options[:use_client_cert]
+        http_conn.cert = @options[:ssl_client_cert]
+        http_conn.key = @options[:ssl_client_key]
+      end
       http_conn.verify_mode = @options[:ssl_verify_mode]
+      http_conn.ssl_version = @options[:ssl_version] if @options[:ssl_version]
       http_conn.read_timeout = @options[:read_timeout]
+      http_conn.ca_file = @options[:ca_file] if @options[:ca_file]
       http_conn
     end
 
     def uri
-      uri = URI.parse(@options[:site])
+      URI.parse(@options[:site])
     end
 
     def authenticated?
@@ -64,6 +72,17 @@ module JIRA
     end
 
     private
+
+    def execute_request(request)
+      add_cookies(request) if options[:use_cookies]
+      request.basic_auth(@options[:username], @options[:password]) if @options[:username] && @options[:password]
+
+      response = basic_auth_http_conn.request(request)
+      @authenticated = response.is_a? Net::HTTPOK
+      store_cookies(response) if options[:use_cookies]
+
+      response
+    end
 
     def request_path(url)
       parsed_uri = URI(url)
@@ -86,7 +105,7 @@ module JIRA
 
     def add_cookies(request)
       cookie_array = @cookies.values.map { |cookie| "#{cookie.name}=#{cookie.value[0]}" }
-      cookie_array +=  Array(@options[:additional_cookies]) if @options.key?(:additional_cookies)
+      cookie_array += Array(@options[:additional_cookies]) if @options.key?(:additional_cookies)
       request.add_field('Cookie', cookie_array.join('; ')) if cookie_array.any?
       request
     end
