@@ -1,6 +1,5 @@
 module JIRA
   module Resource
-
     class IssuelinkFactory < JIRA::BaseFactory # :nodoc:
     end
 
@@ -9,14 +8,13 @@ module JIRA
     class Issue < JIRA::Base; end
 
     class Issuelink < JIRA::Base
-      has_one :type, :class => JIRA::Resource::Issuelinktype
-      has_one :inwardIssue, :class => JIRA::Resource::Issue
-      has_one :outwardIssue, :class => JIRA::Resource::Issue
+      has_one :type, class: JIRA::Resource::Issuelinktype
+      has_one :inwardIssue, class: JIRA::Resource::Issue
+      has_one :outwardIssue, class: JIRA::Resource::Issue
 
       def self.endpoint_name
         'issueLink'
       end
-
     end
   end
 end
