@@ -16,7 +16,7 @@ RSpec.shared_examples 'Client Common Tests' do
   end
 
   it 'merges headers' do
-    expect(subject.send(:merge_default_headers, {})).to eq({'Accept' => 'application/json'})
+    expect(subject.send(:merge_default_headers, {})).to eq('Accept' => 'application/json')
   end
 
   describe 'http methods' do
@@ -30,17 +30,17 @@ RSpec.shared_examples 'Client Common Tests' do
       # response for merging headers for http methods with body
       expect(subject).to receive(:merge_default_headers).exactly(2).times.with(content_type_header)
 
-      [:delete, :get, :head].each { |method| subject.send(method, '/path', {}) }
-      [:post, :put].each {|method| subject.send(method, '/path', '', content_type_header)}
+      %i[delete get head].each { |method| subject.send(method, '/path', {}) }
+      %i[post put].each { |method| subject.send(method, '/path', '', content_type_header) }
     end
 
     it 'calls the generic request method' do
-      [:delete, :get, :head].each do |method|
+      %i[delete get head].each do |method|
         expect(subject).to receive(:request).with(method, '/path', nil, headers).and_return(successful_response)
         subject.send(method, '/path', {})
       end
 
-      [:post, :put].each do |method|
+      %i[post put].each do |method|
         expect(subject).to receive(:request).with(method, '/path', '', merged_headers)
         subject.send(method, '/path', '', {})
       end
@@ -59,17 +59,60 @@ RSpec.shared_examples 'Client Common Tests' do
       expect(subject.Project.find('123')).to eq(find_result)
     end
   end
+
+  describe 'SSL client options' do
+    context 'without certificate and key' do
+      let(:options) { { use_client_cert: true } }
+      subject { JIRA::Client.new(options) }
+
+      it 'raises an ArgumentError' do
+        expect { subject }.to raise_exception(ArgumentError, 'Options: :cert_path or :ssl_client_cert must be set when :use_client_cert is true')
+        options[:ssl_client_cert] = '<cert></cert>'
+        expect { subject }.to raise_exception(ArgumentError, 'Options: :key_path or :ssl_client_key must be set when :use_client_cert is true')
+      end
+    end
+  end
 end
 
 RSpec.shared_examples 'HttpClient tests' do
   it 'makes a valid request' do
-    [:delete, :get, :head].each do |method|
+    %i[delete get head].each do |method|
       expect(subject.request_client).to receive(:make_request).with(method, '/path', nil, headers).and_return(successful_response)
       subject.send(method, '/path', headers)
     end
-    [:post, :put].each do |method|
+    %i[post put].each do |method|
       expect(subject.request_client).to receive(:make_request).with(method, '/path', '', merged_headers).and_return(successful_response)
       subject.send(method, '/path', '', headers)
+    end
+  end
+end
+
+RSpec.shared_examples 'OAuth Common Tests' do
+  include_examples 'Client Common Tests'
+
+  specify { expect(subject.request_client).to be_a JIRA::OauthClient }
+
+  it 'allows setting an access token' do
+    token = double
+    expect(OAuth::AccessToken).to receive(:new).with(subject.consumer, '', '').and_return(token)
+
+    expect(subject.authenticated?).to be_falsey
+    access_token = subject.set_access_token('', '')
+    expect(access_token).to eq(token)
+    expect(subject.access_token).to eq(token)
+    expect(subject.authenticated?).to be_truthy
+  end
+
+  describe 'that call a oauth client' do
+    specify 'which makes a request' do
+      %i[delete get head].each do |method|
+        expect(subject.request_client).to receive(:make_request).with(method, '/path', nil, headers).and_return(successful_response)
+        subject.send(method, '/path', {})
+      end
+      %i[post put].each do |method|
+        expect(subject.request_client).to receive(:make_request).with(method, '/path', '', merged_headers).and_return(successful_response)
+        subject.send(method, '/path', '', {})
+      end
     end
   end
 end
@@ -78,16 +121,16 @@ describe JIRA::Client do
   let(:request) { subject.request_client.class }
   let(:successful_response) do
     response = double('response')
-    allow(response).to receive(:kind_of?).with(Net::HTTPSuccess).and_return(true)
+    allow(response).to receive(:is_a?).with(Net::HTTPSuccess).and_return(true)
     response
   end
-  let(:content_type_header) { {'Content-Type' => 'application/json'} }
-  let(:headers) { {'Accept' => 'application/json'} }
+  let(:content_type_header) { { 'Content-Type' => 'application/json' } }
+  let(:headers) { { 'Accept' => 'application/json' } }
   let(:merged_headers) { headers.merge(content_type_header) }
 
   context 'behaviour that applies to all client classes irrespective of authentication method' do
     it 'allows the overriding of some options' do
-      client = JIRA::Client.new({:consumer_key => 'foo', :consumer_secret => 'bar', :site => 'http://foo.com/'})
+      client = JIRA::Client.new(consumer_key: 'foo', consumer_secret: 'bar', site: 'http://foo.com/')
       expect(client.options[:site]).to eq('http://foo.com/')
       expect(JIRA::Client::DEFAULT_OPTIONS[:site]).not_to eq('http://foo.com/')
     end
@@ -97,11 +140,13 @@ describe JIRA::Client do
     subject { JIRA::Client.new(username: 'foo', password: 'bar', auth_type: :basic) }
 
     before(:each) do
-      stub_request(:get, 'https://foo:bar@localhost:2990/jira/rest/api/2/project')
-        .to_return(status: 200, body: '[]', headers: {} )
+      stub_request(:get, 'https://localhost:2990/jira/rest/api/2/project')
+        .with(headers: { 'Authorization' => "Basic #{Base64.strict_encode64('foo:bar').chomp}" })
+        .to_return(status: 200, body: '[]', headers: {})
 
-      stub_request(:get, 'https://foo:badpassword@localhost:2990/jira/rest/api/2/project').
-        to_return(status: 401, headers: {} )
+      stub_request(:get, 'https://localhost:2990/jira/rest/api/2/project')
+      .with(headers: { 'Authorization' => "Basic #{Base64.strict_encode64('foo:badpassword').chomp}" })
+        .to_return(status: 401, headers: {})
     end
 
     include_examples 'Client Common Tests'
@@ -114,18 +159,17 @@ describe JIRA::Client do
       expect(subject.options[:password]).to eq('bar')
     end
 
-    it 'fails with wrong user name and password' do
-      bad_login = JIRA::Client.new(username: 'foo', password: 'badpassword', auth_type: :basic)
-      expect(bad_login.authenticated?).to be_falsey
-      expect{bad_login.Project.all}.to raise_error JIRA::HTTPError
-    end
-
     it 'only returns a true for #authenticated? once we have requested some data' do
-      expect(subject.authenticated?).to be_falsey
+      expect(subject.authenticated?).to be_nil
       expect(subject.Project.all).to be_empty
       expect(subject.authenticated?).to be_truthy
     end
 
+    it 'fails with wrong user name and password' do
+      bad_login = JIRA::Client.new(username: 'foo', password: 'badpassword', auth_type: :basic)
+      expect(bad_login.authenticated?).to be_falsey
+      expect { bad_login.Project.all }.to raise_error JIRA::HTTPError
+    end
   end
 
   context 'with cookie authentication' do
@@ -134,27 +178,27 @@ describe JIRA::Client do
     let(:session_cookie) { '6E3487971234567896704A9EB4AE501F' }
     let(:session_body) do
       {
-        'session': {'name' => "JSESSIONID", 'value' => session_cookie },
-        'loginInfo': {'failedLoginCount' => 1, 'loginCount' => 2,
-                      'lastFailedLoginTime' => (DateTime.now - 2).iso8601,
-                      'previousLoginTime' => (DateTime.now - 5).iso8601 }
+        'session': { 'name' => 'JSESSIONID', 'value' => session_cookie },
+        'loginInfo': { 'failedLoginCount' => 1, 'loginCount' => 2,
+                       'lastFailedLoginTime' => (DateTime.now - 2).iso8601,
+                       'previousLoginTime' => (DateTime.now - 5).iso8601 }
       }
     end
 
     before(:each) do
       # General case of API call with no authentication, or wrong authentication
-      stub_request(:post, 'https://localhost:2990/jira/rest/auth/1/session').
-        to_return(status: 401, headers: {} )
+      stub_request(:post, 'https://localhost:2990/jira/rest/auth/1/session')
+        .to_return(status: 401, headers: {})
 
       # Now special case of API with correct authentication.  This gets checked first by RSpec.
       stub_request(:post, 'https://localhost:2990/jira/rest/auth/1/session')
         .with(body: '{"username":"foo","password":"bar"}')
         .to_return(status: 200, body: session_body.to_json,
-                   headers: { 'Set-Cookie': "JSESSIONID=#{session_cookie}; Path=/; HttpOnly"})
+                   headers: { 'Set-Cookie': "JSESSIONID=#{session_cookie}; Path=/; HttpOnly" })
 
       stub_request(:get, 'https://localhost:2990/jira/rest/api/2/project')
-        .with(headers: { cookie: "JSESSIONID=#{session_cookie}" } )
-        .to_return(status: 200, body: '[]', headers: {} )
+        .with(headers: { cookie: "JSESSIONID=#{session_cookie}" })
+        .to_return(status: 200, body: '[]', headers: {})
     end
 
     include_examples 'Client Common Tests'
@@ -178,51 +222,72 @@ describe JIRA::Client do
     end
   end
 
-  context 'oath2 authentication' do
-    subject { JIRA::Client.new(consumer_key: 'foo', consumer_secret: 'bar') }
+  context 'with jwt authentication' do
+    subject do
+      JIRA::Client.new(
+        issuer: 'foo',
+        base_url: 'https://host.tld',
+        shared_secret: 'shared_secret_key',
+        auth_type: :jwt
+      )
+    end
+
+    before(:each) do
+      stub_request(:get, 'https://localhost:2990/jira/rest/api/2/project')
+          .with(headers: {"Authorization" => /JWT .+/})
+          .to_return(status: 200, body: '[]', headers: {})
+    end
 
     include_examples 'Client Common Tests'
+    include_examples 'HttpClient tests'
 
-    specify { expect(subject.request_client).to be_a JIRA::OauthClient }
+    specify { expect(subject.request_client).to be_a JIRA::JwtClient }
 
-    it 'allows setting an access token' do
-      token = double
-      expect(OAuth::AccessToken).to receive(:new).with(subject.consumer, 'foo', 'bar').and_return(token)
+    it 'sets the username and password' do
+      expect(subject.options[:shared_secret]).to eq('shared_secret_key')
+    end
 
+    context 'with a incorrect jwt key' do
+      before do
+        stub_request(:get, 'https://localhost:2990/jira/rest/api/2/project')
+            .with(headers: {"Authorization" => /JWT .+/})
+            .to_return(status: 401, body: '[]', headers: {})
+      end
+
+      it 'is not authenticated' do
+        expect(subject.authenticated?).to be_falsey
+      end
+
+      it 'raises a JIRA::HTTPError when trying to fetch projects' do
+        expect { subject.Project.all }.to raise_error JIRA::HTTPError
+      end
+    end
+
+    it 'only returns a true for #authenticated? once we have requested some data' do
       expect(subject.authenticated?).to be_falsey
-      access_token = subject.set_access_token('foo', 'bar')
-      expect(access_token).to eq(token)
-      expect(subject.access_token).to eq(token)
+      expect(subject.Project.all).to be_empty
       expect(subject.authenticated?).to be_truthy
     end
+  end
 
-    it 'allows initializing the access token' do
-      request_token = OAuth::RequestToken.new(subject.consumer)
-      allow(subject.consumer).to receive(:get_request_token).and_return(request_token)
-      mock_access_token = double
-      expect(request_token).to receive(:get_access_token).with(:oauth_verifier => 'abc123').and_return(mock_access_token)
-      subject.init_access_token(:oauth_verifier => 'abc123')
-      expect(subject.access_token).to eq(mock_access_token)
-    end
+  context 'oauth authentication' do
+    subject { JIRA::Client.new(consumer_key: 'foo', consumer_secret: 'bar') }
 
-    specify 'that has specific default options' do
-      [:signature_method, :private_key_file].each do |key|
-        expect(subject.options[key]).to eq(JIRA::Client::DEFAULT_OPTIONS[key])
-      end
-    end
+    include_examples 'OAuth Common Tests'
+  end
 
-    describe 'that call a oauth client' do
-      specify 'which makes a request' do
-        [:delete, :get, :head].each do |method|
-          expect(subject.request_client).to receive(:make_request).with(method, '/path', nil, headers).and_return(successful_response)
-          subject.send(method, '/path', {})
-        end
-        [:post, :put].each do |method|
-          expect(subject.request_client).to receive(:make_request).with(method, '/path', '', merged_headers).and_return(successful_response)
-          subject.send(method, '/path', '', {})
-        end
-      end
+  context 'with oauth_2legged' do
+    subject { JIRA::Client.new(consumer_key: 'foo', consumer_secret: 'bar', auth_type: :oauth_2legged) }
+
+    include_examples 'OAuth Common Tests'
+  end
+
+  context 'with unknown options' do
+    let(:options) { { 'username' => 'foo', 'password' => 'bar', auth_type: :basic } }
+    subject { JIRA::Client.new(options) }
+
+    it 'raises an ArgumentError' do
+      expect { subject }.to raise_exception(ArgumentError, 'Unknown option(s) given: ["username", "password"]')
     end
   end
 end
-
