@@ -99,7 +99,16 @@ client.Issue.jql(a_normal_jql_search, max_results: 500)
 # # Create an issue
 # # ---------------
 # issue = client.Issue.build
-# issue.save({"fields"=>{"summary"=>"blarg from in example.rb","project"=>{"id"=>"10001"},"issuetype"=>{"id"=>"3"}}})
+# labels = ['label1', 'label2']
+# issue.save({
+#   "fields" => {
+#     "summary"   => "blarg from in example.rb",
+#     "project"   => {"key" => "SAMPLEPROJECT"},
+#     "issuetype" => {"id" => "3"},
+#     "labels"    => labels,
+#     "priority"  => {"id" => "1"}
+#   }
+# })
 # issue.fetch
 # pp issue
 #
@@ -156,6 +165,14 @@ client.Issue.jql(a_normal_jql_search, max_results: 500)
 # # Update an existing comment
 # # --------------------------
 # issue.comments.first.save({"body" => "an updated comment frome example.rb"})
+
+
+# # Add attachment to Issue
+# # ------------------------
+#  issue = client.Issue.find('PROJ-1')
+#  attachment = issue.attachments.build
+#  attachment.save('file': '/path/to/file')
+#
 
 # List all available link types
 # ------------------------------
