@@ -2,14 +2,16 @@ require 'cgi'
 
 module JIRA
   module Resource
-
     class AgileFactory < JIRA::BaseFactory # :nodoc:
     end
 
     class Agile < JIRA::Base
-
-      def self.all(client)
-        response = client.get(path_base(client) + '/board')
+      # @param client [JIRA::Client]
+      # @param options [Hash<Symbol, Object>]
+      # @return [Hash]
+      def self.all(client, options = {})
+        opts = options.empty? ? '' : "?#{hash_to_query_string(options)}"
+        response = client.get(path_base(client) + "/board#{opts}")
         parse_json(response.body)
       end
 
@@ -23,9 +25,10 @@ module JIRA
         response = client.get(path_base(client) + "/board/#{board_id}/issue?#{hash_to_query_string(options)}")
         json = parse_json(response.body)
         # To get Issue objects with the same structure as for Issue.all
-        issue_ids = json['issues'].map { |issue|
+        return {} if json['issues'].size.zero?
+        issue_ids = json['issues'].map do |issue|
           issue['id']
-        }
+        end
         client.Issue.jql("id IN(#{issue_ids.join(', ')})")
       end
 
@@ -41,7 +44,7 @@ module JIRA
         parse_json(response.body)
       end
 
-      def self.get_projects_full(client, board_id, options = {})
+      def self.get_projects_full(client, board_id, _options = {})
         response = client.get(path_base(client) + "/board/#{board_id}/project/full")
         parse_json(response.body)
       end
@@ -72,6 +75,5 @@ module JIRA
         self.class.path_base(client)
       end
     end
-
   end
 end
