@@ -1,10 +1,8 @@
 module JIRA
   module Resource
-
     class ComponentFactory < JIRA::BaseFactory # :nodoc:
     end
 
-    class Component < JIRA::Base ; end
-
+    class Component < JIRA::Base; end
   end
 end
