@@ -1,10 +1,8 @@
 module JIRA
   module Resource
-
     class IssuetypeFactory < JIRA::BaseFactory # :nodoc:
     end
 
-    class Issuetype < JIRA::Base ; end
-
+    class Issuetype < JIRA::Base; end
   end
 end
