@@ -1,10 +1,8 @@
 module JIRA
   module Resource
-
     class PriorityFactory < JIRA::BaseFactory # :nodoc:
     end
 
-    class Priority < JIRA::Base ; end
-
+    class Priority < JIRA::Base; end
   end
 end
