@@ -1,10 +1,8 @@
 module JIRA
   module Resource
-
     class ResolutionFactory < JIRA::BaseFactory # :nodoc:
     end
 
-    class Resolution < JIRA::Base ; end
-
+    class Resolution < JIRA::Base; end
   end
 end
