@@ -1,6 +1,5 @@
 module JIRA
   module Resource
-
     class ServerInfoFactory < JIRA::BaseFactory # :nodoc:
     end
 
@@ -12,7 +11,7 @@ module JIRA
       def self.all(client, options = {})
         response = client.get(collection_path(client))
         json = parse_json(response.body)
-        self.new(client, {:attrs => json}.merge(options))
+        new(client, { attrs: json }.merge(options))
       end
     end
   end
