@@ -1,10 +1,8 @@
 module JIRA
   module Resource
-
     class VersionFactory < JIRA::BaseFactory # :nodoc:
     end
 
-    class Version < JIRA::Base ; end
-
+    class Version < JIRA::Base; end
   end
 end
