@@ -1,3 +1,3 @@
 module JIRA
-  VERSION = '1.4.3'
+  VERSION = '2.3.0'.freeze
 end
